Refactor: update MemberControls, Extract TotalReputation
