+++ src/components/common/ColonyMembers/ColonyMembers.css
@@ -29,28 +29,6 @@
   position: relative;
 }
 
-.controls li {
-  width: fit-content;
-}
-
-.controls li + li {
-  margin: 18px 0 0;
-}
-
-.teamReputationPointsContainer {
-  margin-bottom: 50px;
-}
-
-.teamReputationPointsContainer h4 {
-  font-size: var(--size-smallish);
-  line-height: 18px;
-}
-
-.reputationPoints {
-  line-height: 18px;
-  color: var(--dark);
-}
-
 @media screen and query700 {
   .main {
     height: 100%;
