+++ src/components/common/ColonyMembers/MemberControls/MemberControls.css
@@ -1,0 +1,7 @@
+.controls li {
+  width: fit-content;
+}
+
+.controls li + li {
+  margin: 18px 0 0;
+}
