+++ src/components/common/ColonyMembers/TotalReputation/TotalReputation.css
@@ -1,0 +1,13 @@
+.teamReputationPointsContainer {
+  margin-bottom: 50px;
+}
+
+.teamReputationPointsContainer h4 {
+  font-size: var(--size-smallish);
+  line-height: 18px;
+}
+
+.reputationPoints {
+  line-height: 18px;
+  color: var(--dark);
+}
