+++ src/components/common/ColonyMembers/TotalReputation/TotalReputation.css.d.ts
@@ -1,0 +1,13 @@
+declare namespace TotalReputationCssNamespace {
+  export interface ITotalReputationCss {
+    reputationPoints: string;
+    teamReputationPointsContainer: string;
+  }
+}
+
+declare const TotalReputationCssModule: TotalReputationCssNamespace.ITotalReputationCss & {
+  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
+  locals: TotalReputationCssNamespace.ITotalReputationCss;
+};
+
+export = TotalReputationCssModule;
